requirements.txt
streamlit==1.23.1
plotly==5.18.0
pandas==1.5.3
numpy==1.21.5
folium==0.15.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
haversine==2.8.0
streamlit-folium==0.13.0
Pillow==9.3.0